pyproject.toml
[project]
name = "low-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.40.0",
]
